Move selectors to their respective folders
