+++ user-registry-ui/ethereum/index.js
@@ -1,1 +1,2 @@
 export { ethereumReducer } from './reducer'
+export * from './selectors'
