+++ user-registry-ui/ethereum/selectors.js
@@ -1,2 +1,1 @@
-export const getGithubUsername = state => state.github.username
 export const getEthereumAccount = state => state.ethereum.account
