+++ user-registry-ui/ethereum/selectors.test.js
@@ -1,0 +1,6 @@
+import { getEthereumAccount } from './selectors'
+
+it('selects the ethereum account', () => {
+  const account = 'some ethereum account'
+  expect(getEthereumAccount({ ethereum: { account } })).toEqual(account)
+})
