+++ user-registry-ui/github/index.js
@@ -1,1 +1,2 @@
 export { githubReducer } from './reducer'
+export * from './selectors'
