+++ user-registry-ui/github/selectors.js
@@ -1,0 +1,1 @@
+export const getGithubUsername = state => state.github.username
