+++ user-registry-ui/github/selectors.test.js
@@ -1,0 +1,6 @@
+import { getGithubUsername } from './selectors'
+
+it('selects the github username', () => {
+  const username = 'some user name'
+  expect(getGithubUsername({ github: { username } })).toEqual(username)
+})
